Fix ContextModel memory bug, add Docker build env
 - Superficial changes to get the project to compile with GCC: added
missing STL #inlcudes etc.

 - Added platform-dependent preamble to the SConscript to get scons to
compile with g++ if we're on Linux. If we're on OS X we still compile
with clang/LLVM.

 - Added Dockerfile to create a containerised Linux build environment in
which we can run valgrind to check for leaks (Apple broke valgrind with
macOS 10.12).

 - Added more support for the short-term model in the MVS. This will be
fully integrated and tested in the next (few) commit(s).

 - Fixed a nasty bug whereby we weren't making deep copies of TreeNodes
(and thus ContextModels, SequneceModels and Viewpoints). This caused a
problem whereby Viewpoints for the long-term model would be duplicated
for the short-term model and subsequently reset, causing the long-term
model to be wiped out (and a subsequent segfault).