Milestone: implemented basic linked viewpoints
 - Implementation of linked viewpoints modelling (pointwise) basic types
complete. There is a slight asymmetry in the implementation in that a
viewpoint linking types T and S must be instantiated twice in order to
predict both T and S (namely one instance to predict T and another to
predict S). This is to avoid some multiple inheritance yuckiness that
would otherwise have to result. Obviously this isn't super efficient but
that's not a concern right now.

 - Build script updated to support both debug and release builds (using
the BUILD_TYPE argument). With the new linked viewpoints we need all the
help we can get w.r.t. optimisations.

 - Evaluation code now has slighlty more helpful output if you're
waiting around for it to finish.