improved linked VPs + implemented posinbar
 - Added better error checking to ChoraleRest: can now only be
instantiated with rests that are a crotchet multiple in length.

 - Added time signatures to ChoraleEvent and in all the other necessary
places.

 - Implemented and written (passing) tests for posinbar. This will be
used to implemented a "metre" viewpoint (triply-linked VP), hopefully in
the next commit.

 - Even types that have different event stream lengths (such as intref
and seqint) can now be linked using the new general VP linking scheme.
This is achieved under the assumption that the VP with a shorter event
stream is a first-order difference VP (such as seqint) and aligning the
two event streams appropriately before zipping them together (see
EventPair::zip_tail).